MS CRM Document Scanner
This JavaScript, adds a new button on MS CRM entities you are
interested so documents can be directly scanned into entity document
library.
It also can be used to show a thumbnail of your images right in your
main form.

Special accesses will be required for IE to be able to run this
JavaScript.